requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-filter==23.5
django-rest==0.8.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.1
